+++ src/dapps/ManageENS/supportedTxt.js
@@ -11,25 +11,31 @@ const isUrl = function (input) {
 };
 
 const isEmail = function (input) {
-  if (!input) return false;
+  if (!input || input === '') return false;
   const atIndex = input.indexOf('@');
-  const parsedEmailName = normalise(input.substr(0, atIndex));
-  const parsedEmailHost = normalise(input.substr(atIndex + 1, input.length));
   const emailRegex = new RegExp(
     // eslint-disable-next-line no-useless-escape
     /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,24}))$/
   );
-  return emailRegex.test(`${parsedEmailName}@${parsedEmailHost}`.toLowerCase());
+
+  try {
+    const parsedEmailName = normalise(input.substr(0, atIndex));
+    const parsedEmailHost = normalise(input.substr(atIndex + 1, input.length));
+    return emailRegex.test(
+      `${parsedEmailName}@${parsedEmailHost}`.toLowerCase()
+    );
+  } catch (e) {
+    return emailRegex.test(input);
+  }
 };
 
 const isString = function (input) {
-  if (!input) return false;
-  const parsedInput = normalise(input);
-  return typeof parsedInput === 'string';
+  if (!input || input === '') return false;
+  return typeof input === 'string';
 };
 
 const isHandle = function (input) {
-  if (!input) return false;
+  if (!input || input === '') return false;
   const atIndex = input.indexOf('@');
   const parsedInput = normalise(input.substr(atIndex + 1, input.length));
   if (!isString(parsedInput)) return false;