+++ src/dapps/index.js
@@ -62,16 +62,6 @@ const dapps = {
     lastUpdated: '08/14/2019',
     releaseDate: '06/14/2019'
   },
-  makerOld: {
-    route: '/interface/dapps/maker-sai',
-    icon: makerdai,
-    iconDisabled: makerdaiDisabled,
-    title: 'dappsMaker.maker_title_old',
-    desc: 'dappsMaker.maker-shutdown',
-    supportedNetworks: [ETH.name, KOV.name],
-    lastUpdated: '08/14/2019',
-    releaseDate: '06/14/2019'
-  },
   ambrpay: {
     route: '/interface/dapps/ambrpay',
     icon: ambrpay,