+++ src/globalErrorHandler.js
@@ -2,7 +2,8 @@ import { Toast } from '@/helpers';
 import Vue from 'vue';
 const ERRORS = {
   'The operation is insecure.': 'errorsGlobal.browser-not-supported',
-  'Access denied.': 'errorsGlobal.access-denied'
+  'Access denied.': 'errorsGlobal.access-denied',
+  'UNKNOWN_ERROR (0x6804)': 'ledgerError.unknown-0x6804'
 };
 const WARNING = {
   'Network Error': 'errorsGlobal.network-error',
@@ -16,11 +17,11 @@ export default event => {
   const errorValues = Object.keys(ERRORS);
   const warningValues = Object.keys(WARNING);
   const foundError = errorValues.find(item => {
-    return event.value.includes(item);
+    return event.value && event.value.includes(item);
   });
 
   const foundWarning = warningValues.find(item => {
-    return event.value.includes(item);
+    return event.value && event.value.includes(item);
   });
 
   if (foundError) {