+++ src/networks/types/FUSE.js
@@ -1,0 +1,14 @@
+import fuse from '@/assets/images/networks/fuse.svg';
+
+export default {
+  name: 'FUSE',
+  name_long: 'Fuse',
+  homePage: 'https://fuse.io/',
+  blockExplorerTX: 'https://explorer.fuse.io/tx/[[txHash]]',
+  blockExplorerAddr: 'https://explorer.fuse.io/address/[[address]]',
+  chainID: 122,
+  tokens: [],
+  contracts: [],
+  icon: fuse,
+  currencyName: 'FUSE'
+};