+++ src/store/main/state.js
@@ -58,7 +58,8 @@ const state = {
   linkQuery: {},
   addressBook: addressBook,
   locale: store.get('locale') !== undefined ? store.get('locale') : 'en_US',
-  tempHide: false
+  tempHide: false,
+  gasLimitWarning: 100
 };
 
 export default state;