+++ src/wallets/hardware/trezor/errorHandler.js
@@ -11,11 +11,14 @@ const ERRORS = {
   'Transport is missing': 'trezorError.transport-missing',
   'EthAppPleaseEnableContractData: Please enable Contract data on the Ethereum app Settings':
     'trezorError.turn-on-contract-data',
+  'Initialize failed: Response of unexpected type: Address. Should be Features':
+    'trezor.initializing-failed',
   Cancelled: 'trezorError.cancelled',
   'Iframe timeout': 'trezor.iframe-timeout',
   'Browser not supported': 'trezor.unsupported-browser',
   'popup failed to open': 'trezor.popup-failed-to-open',
-  'Safety check failed': 'trezor.safety-check-failed'
+  'Safety check failed': 'trezor.safety-check-failed',
+  'TrezorConnect not yet initialized': 'trezor.trezor-connect-not-initialized'
 };
 
 const WARNING = {};