+++ tests/unit/specs/src/partners/changelly/changelly.spec.js
@@ -1,7 +1,12 @@
 //import Changelly from '@/partners/changelly/changelly.js';
-
+import changellyCalls from '@/partners/changelly/changelly-calls';
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
 describe('changelly.js', () => {
+  it('not cancelled', async done => {
+    const val = await changellyCalls.getStatus('u04b5b8bx9h7x3x5', 'ETH');
+    console.log(val); // todo remove dev item
+    done();
+  });
   /*xit('testing', async () => {
     const fromCurrency = currency.symbol;
     const toCurrency = 'ETH';